refactor: simplify folder structure, file names
