+++ product-service/src/functions/getProductsList.ts
@@ -1,7 +1,7 @@
 import { handlerPath } from '@libs/handler-resolver';
 
 export default {
-  handler: `${handlerPath(__dirname)}/lambda/getProductsListLambda.main`,
+  handler: `${handlerPath(__dirname)}/lambda/getProductsList.main`,
   events: [
     {
       http: {
