+++ product-service/src/functions/index.ts
@@ -1,2 +1,2 @@
-export { default as products } from './products/getProductsList';
-export { default as productById } from './products/getProductById';
+export { default as products } from './getProductsList';
+export { default as productById } from './getProductById';
