+++ product-service/src/utils/schema.ts
